Refactor: 업로드 버튼 이름 및 children으로 수정
