+++ components/reusable/Buttons/SignoutBtn.tsx
@@ -1,6 +1,6 @@
 import { styled } from 'styled-components';
 import theme from '../../../styles/theme';
-import { ButtonProps } from './UploadBtn';
+import { ButtonProps } from './SquareBtn';
 
 const SignoutBtn = ({ onClick }: ButtonProps) => {
   return <ButtonWrapper onClick={onClick}>탈퇴하기</ButtonWrapper>;
